+++ Assets/Scripts/SS3D/Systems/PlayerControl/PlayerControlSystem.cs
@@ -23,10 +23,8 @@ namespace SS3D.Systems.PlayerControl
         [SerializeField] private NetworkObject _unauthorizedUserPrefab;
         [SerializeField] private NetworkObject _soulPrefab;
 
-        [SyncObject]
-        private readonly SyncList<Soul> _serverSouls = new();
-        [SyncObject] 
-        private readonly SyncList<Soul> _onlineSouls = new();
+        [SyncObject] public readonly SyncList<Soul> ServerSouls = new();
+        [SyncObject] public readonly SyncList<Soul> OnlineSouls = new();
 
         protected override void OnStart()
         {
@@ -46,8 +44,8 @@ namespace SS3D.Systems.PlayerControl
 
         private void AddEventListeners()
         {
-            _serverSouls.OnChange += HandleServerSoulsChanged;
-            _onlineSouls.OnChange += HandleOnlineSouls;
+            ServerSouls.OnChange += HandleServerSoulsChanged;
+            OnlineSouls.OnChange += HandleOnlineSouls;
         }
 
         private void HandleRemoteConnectionState(NetworkConnection conn, RemoteConnectionStateArgs remoteConnectionStateArgs)
@@ -69,7 +67,7 @@ namespace SS3D.Systems.PlayerControl
 
             Soul soul = changeType == ChangeType.Addition ? newItem : oldItem;
 
-            OnlineSoulsChanged serverSoulsChanged = new(_onlineSouls.ToList(), changeType, soul);
+            OnlineSoulsChanged serverSoulsChanged = new(OnlineSouls.ToList(), changeType, soul);
             serverSoulsChanged.Invoke(this);
         }
 
@@ -79,7 +77,7 @@ namespace SS3D.Systems.PlayerControl
 
             Soul soul = changeType == ChangeType.Addition ? newItem : oldItem;
 
-            ServerSoulsChanged serverSoulsChanged = new(_serverSouls.ToList(), changeType, soul);
+            ServerSoulsChanged serverSoulsChanged = new(ServerSouls.ToList(), changeType, soul);
             serverSoulsChanged.Invoke(this);
         }
 
@@ -117,7 +115,7 @@ namespace SS3D.Systems.PlayerControl
                 match = Instantiate(_soulPrefab).GetComponent<Soul>();
                 match.UpdateCkey(string.Empty, ckey, true);
 
-                _serverSouls.Add(match);
+                ServerSouls.Add(match);
 
                 ServerManager.Spawn(match.gameObject);
             }
@@ -129,7 +127,7 @@ namespace SS3D.Systems.PlayerControl
             NetworkObject networkObject = match.NetworkObject;
             networkObject.GiveOwnership(conn);
             
-            _onlineSouls.Add(match);
+            OnlineSouls.Add(match);
         }
 
         [Server]
@@ -152,7 +150,7 @@ namespace SS3D.Systems.PlayerControl
                 Soul soul = networkIdentity.GetComponent<Soul>();
                 if (soul != null)
                 {
-                    _onlineSouls.Remove(soul);
+                    OnlineSouls.Remove(soul);
                     soul.RemoveOwnership();
                     return;
                 }
@@ -164,17 +162,17 @@ namespace SS3D.Systems.PlayerControl
 
         public string GetCkey(NetworkConnection conn)
         {
-            return _serverSouls.Find(soul => soul.Owner == conn)?.Ckey;
+            return ServerSouls.Find(soul => soul.Owner == conn)?.Ckey;
         }
 
         public Soul GetSoul(string ckey)
         {
-            return _serverSouls.Find(soul => soul.Ckey == ckey);
+            return ServerSouls.Find(soul => soul.Ckey == ckey);
         }
 
         public Soul GetSoul(NetworkConnection conn)
         {
-            return _serverSouls.Find(soul => soul.Owner == conn);
+            return ServerSouls.Find(soul => soul.Owner == conn);
         }
     }
 }
