--- Assets/Scripts/SS3D/UI/IngameConsole.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dcac00fbb0b668242bb98c08361e345a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
